refactor: Extract transport creation logic into reusable helper function
